Allow raw and smart pointers as base return types for request handler functions
Request with `std::shared_ptr`, `std::unique_ptr` and raw pointer values for
`_RETURN_TYPE_` do not have the handler function return type wrapped in a
`std::optional` and so `nullptr` is returned when no objects are subscribed
to the request.

The same is true for requests with `std::optional` values for `_RETURN_TYPE_`
for cases where request struct declarations need to be explicit.